Move network scan to network file.
